Update patch set 2
Patch Set 2:

Don't worry Jean Baptiste, this is a really minor contribution, it's normal that you don't loose too much time with it. Besides, it allow me to learn how the repo tool works :). 
Now that Romain is added to the reviewer... sounds that there is a big proportion of frenches on this patch (i'm french too ;) )

Patch-set: 2