Update patch set 1
Patch Set 1: Do not submit

(1 inline comment)

The problem you describe only happens with ArtistAlbumBrowserActivity. The other activities are unaffected (though I understand why you might think the problem exists there too). Since this is actually a bug in CursorTreeAdapter, it should be fixed there, not worked around in the app. We have a fix for this in our internal tree already.
Also, I don't think JNI global references are really an issue here. The message you saw in the log is informative only, it does not indicate a permanent leak, just that the cursor was still open when it was garbage collected.

Patch-set: 1
Label: Code-Review=-2